pyproject.toml
[project]
name = "master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.5",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.12",
    "langchain-core==1.2.19",
    "langchain-ollama==1.0.1",
    "langgraph==1.1.2",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.11",
    "langsmith==0.7.17",
    "markdown==3.10.2",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "ollama==0.6.1",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "pathlib==1.0.1",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.3.3",
    "smmap==5.0.3",
    "tenacity==9.1.4",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
]

[project.scripts]
explain_codebase = "explain_codebase.__main__:main"

[dependency-groups]
dev = [
    "black>=26.3.1",
    "ruff>=0.15.6",
]
